--- a/src/templates/functions/finalize.jl
+++ b/src/templates/functions/finalize.jl
@@ -7,11 +7,13 @@ function _build_template_function_finalize(template::CoreTemplate)
     # Get code from "finalize" and remove trailing newline.
     code = chomp(template.yaml["functions"]["finalize"])
 
+    rgf_id = Tuple(reinterpret(UInt32, SHA.sha1(chomp(code))))
+
     # Parse the code into an expression.
     code_ex = Meta.parse("""begin\n$(code)\nend"""; filename="$(template.name).iesopt.template.yaml")
 
     # Convert into a proper function.
-    template.functions[:finalize] = @RuntimeGeneratedFunction(
+    template.functions[:finalize] = _compile_rgf(
         :(function (__virtual__::Virtual)
             __template_name__ = $(template).name
             __parameters__ = __virtual__._parameters
@@ -27,7 +29,8 @@ function _build_template_function_finalize(template::CoreTemplate)
                 rethrow(e)
             end
             return nothing
-        end)
+        end);
+        id=rgf_id,
     )
 
     return nothing